comfirm component refactor
