+++ src/components/Auth/RequestPasswordReset/requestPasswordReset.module.scss
@@ -1,2 +1,2 @@
-@import '../../../sass/utils/container';
-@import '../../../sass/utils/mixins';
+@import '@/sass/utils/container';
+@import '@/sass/utils/mixins';
