+++ src/components/Auth/commonAuthStyles.module.scss
@@ -100,7 +100,7 @@
   text-decoration-thickness: 1px;
 }
 
-// для реєстрації, reset password, confirm component
+// для реєстрації, reset password
 .resetBtn {
   cursor: pointer;
   background-color: transparent;
